Medificar profile 100% funcional

+++ api/src/routes/user.js
@@ -75,8 +75,12 @@ server.post('/user' ,async (req, res, next) => {
 //===============================================
 server.put('/user/:id', (req, res, next) => {
 
-    if (req.user.role !== 'Admin' || req.user.id !== req.params.id) return res.send('<h1>Unauthorized</h1>')
+    if(!req.isAuthenticated()) return res.sendStatus(401)
+
     const { id } = req.params;
+
+    console.log(id)
+    console.log(req.body)
 
     const { name, lastname, email, address, phone, birthdate} = req.body;
 
@@ -87,13 +91,14 @@ server.put('/user/:id', (req, res, next) => {
         address,
         phone,
         birthdate,
-        image
     }, {
         where: {
             id: id
         }
     }).then(modified => {
+        console.log(modified)
         if (modified[0] === 0) {
+            console.log('llegue aca??')
             return res.status(404).send('Usuario no encontrado')
         }
         res.status(200).send('Usuario modificado con exito')
@@ -105,6 +110,9 @@ server.put('/user/:id', (req, res, next) => {
 //===============================================
 
 server.post('/user/:id/image', (req, res, next) => {
+
+    if(!req.isAuthenticated()) return res.sendStatus(401)
+
     const { id } = req.params;
     let { image } = req.body;
 
@@ -128,6 +136,10 @@ server.post('/user/:id/image', (req, res, next) => {
 //  Ruta para eliminar usuario
 //==============================================
 server.delete('/user/:id'/*,[authenticateToken, auths]*/, (req, res, next) => {
+
+    if(!req.isAuthenticated()) return res.sendStatus(401)
+    if(!req.user.role === 'Cliente') return res.sendStatus(401)
+
     const { id } = req.params;
     User.destroy({
         where: {
@@ -166,11 +178,14 @@ server.get('/users', (req, res) => {
 //  Ruta para encontrar resetear la contraseña
 //===================================================
 server.post('/users/:id/passwordReset', async (req, res) => {
+
+    console.log(req.isAuthenticated())
+
     const { id } = req.params;
     const { password } = req.body;
-    
+
     if (password.length < 8) return res.send(400).send('Password no cumple requisitos')
-    
+   
     try {
         const usuario = await User.findOne({ where: {id}})
         const hashedPassword = await bcrypt.hash(password, 9)
